+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- Field type changed to Django JSONField, to be compatible with Django 3.1, 3.2 and 4
 
 ## [0.3.0] - 2021-01-06